Flytt tabellsorteringsutils ut av journalpostmappen
